refactor: drain rts-napi __RTS_* + this_slot; record campaign in the spec
Closes every group in docs/specs/no-mangle-drain.md §6 except rts-egui.

rts-napi — the §2 carve-out held up

  Re-checked before touching anything: 157 of the 162 declarations are the
  `napi_*` / `node_api_*` C contract a compiled `.node` addon links against. The
  name IS the interface there, so they stay hand-written, exactly as §2 says.
  Of the 5 `__RTS_*`, `__RTS_FN_NS_NAPI_LOAD_ADDON` and
  `__RTS_FN_RT_NAPI_NEW_INSTANCE` converted; the other two
  (`INVOKE_METHOD`, `DISPATCH_CALLBACK`) carry `*mut` out-params and are now
  documented carve-outs alongside `ta_view_base_len` and `PROXY_RESOLVE`.

  LOAD_ADDON's `(*const u8, i64)` became `(u64, i64)`, which moved its unit
  tests and `tests/loader_integration.rs` — the §4 rename trap once more.

rts-engine — the `this_slot` trio

  Converted, NOT deleted, though measurement says it is dead: `this_slot` is
  WRITE-ONLY. `rts-primitives/src/function/ops.rs` pushes and pops the
  thread-local `this` stack in seven places and nothing anywhere reads it —
  `__RTS_FN_RT_THIS_GET` has no caller. Unpicking the `pushed_this_slot` control
  flow in the hot dispatch path is its own change with its own verification, so
  it is written down in the spec (§6b) instead of being done as a drain side
  effect.

Spec rewritten to match the tree

  §1 now carries BOTH the before and after tables. New §2b enumerates the seven
  carve-out reasons (out-params; raw bytes that must not gain UTF-8 validation;
  `usize` returns; the `i8` return that is a FLAG not a carve-out), the
  `u64`-address idiom and when `&str` is the wrong answer, and the fact that a
  `macro_rules!` body is invisible to the baker — a real limitation, now stated
  rather than an unexamined silence. §5 records the measured clean-main baseline
  and corrects the link-gap note: it affects every crate below `rts-runtime`
  (rts-shared 36, rts-primitives 35, rts-napi 36), not only rts-std/rts-node.
  §6 marks each group done and promotes the symbol-set diff to a standing
  verification step. §6b records the 27 dead symbols the audit surfaced.

Campaign total: 1126 -> 233 attributes. Excluding the N-API contract (157),
rts-egui (57, untouched — the MANDATORY egui-plan rule requires reading the
frozen plan first, which this work did not do) and rts-macro's generator source
(6), THIRTEEN hand-written declarations remain, each with its reason in the
source.

Verification (spec §5)

  Symbol-set diff: 2191 before, 2191 after, zero added, zero removed.
  cargo check --workspace                    clean
  cargo build --release                      clean
  target/release/rts.exe test                2837/2849 tests, 771/774 files
  cargo test --release -p rts-engine --lib   45 passed
  cargo run -p rts-symbol-baker -- --check   clean
  bash scripts/read_before_commit.sh          no hard violation

Failing set byte-identical to the clean-main baseline. `promise_microtask_order`
reported flaky (passed on serial retry) — a scheduling flake in the harness, not
a failure, and unrelated to symbol declarations.

+++ crates/rts-napi/tests/loader_integration.rs
@@ -89,7 +89,7 @@ fn load_real_addon_returns_valid_exports_handle() {
 
     let path_str = node_path.to_string_lossy().to_string();
     let handle = unsafe {
-        rts_napi::loader::__RTS_FN_NS_NAPI_LOAD_ADDON(path_str.as_ptr(), path_str.len() as i64)
+        rts_napi::loader::__RTS_FN_NS_NAPI_LOAD_ADDON(path_str.as_ptr() as u64, path_str.len() as i64)
     };
 
     assert_ne!(handle, 0, "loader deve devolver um handle de exports não-nulo");
@@ -104,7 +104,7 @@ fn load_real_addon_returns_valid_exports_handle() {
 
     // Idempotência: carregar de novo o mesmo path devolve o MESMO handle.
     let handle2 = unsafe {
-        rts_napi::loader::__RTS_FN_NS_NAPI_LOAD_ADDON(path_str.as_ptr(), path_str.len() as i64)
+        rts_napi::loader::__RTS_FN_NS_NAPI_LOAD_ADDON(path_str.as_ptr() as u64, path_str.len() as i64)
     };
     assert_eq!(handle, handle2, "mesmo path deve ser idempotente (cache)");
